# query: Fixed deduplication inconsistency with vertically deduplicated data. Series without replica label is now assumed part of replication.