--- a/pkg/testutil/testutil.go
+++ b/pkg/testutil/testutil.go
@@ -76,13 +76,20 @@ func Equals(tb testing.TB, exp, act interface{}, v ...interface{}) {
 	if len(v) > 0 {
 		msg = fmt.Sprintf(v[0].(string), v[1:]...)
 	}
-	tb.Fatal(sprintfWithLimit("\033[31m%s:%d:"+msg+"\n\n\texp: %#v\n\n\tgot: %#v%s\033[39m\n\n", filepath.Base(file), line, exp, act, diff(exp, act)))
+	tb.Fatal(sprintfWithLimit(
+		"\033[31m%s:%d:"+msg+"\n\n\texp: "+
+			sprintfWithLimit("%#v", exp)+
+			"\n\n\tgot: "+
+			sprintfWithLimit("%#v", act)+
+			sprintfWithLimit("%s", diff(exp, act))+
+			"\033[39m\n\n", filepath.Base(file), line),
+	)
 }
 
 func sprintfWithLimit(act string, v ...interface{}) string {
 	s := fmt.Sprintf(act, v...)
 	if len(s) > 10000 {
-		return s[:10000] + "...(output trimmed)"
+		return s[:1000] + "...(output trimmed)..." + s[len(s)-1000:]
 	}
 	return s
 }