Fix: https://github.com/thanos-io/thanos/discussions/4274

This PR changes how we treat series without replica (e.g after vertical deduplication). Right now we assume it's another replica allowing us to treat those uniformly for cases when PromQL see both non vertically deduped data and deduped ones.

+ benchmarks. Initial:

```
GOROOT=/home/bwplotka/.gvm/gos/go1.16.3 #gosetup
GOPATH=/home/bwplotka/Repos/thanosgopath #gosetup
/home/bwplotka/.gvm/gos/go1.16.3/bin/go test -c -o /tmp/___BenchmarkQuerySelect_in_github_com_thanos_io_thanos_pkg_query.test github.com/thanos-io/thanos/pkg/query #gosetup
/tmp/___BenchmarkQuerySelect_in_github_com_thanos_io_thanos_pkg_query.test -test.v -test.paniconexit0 -test.bench ^\QBenchmarkQuerySelect\E$ -test.run ^$ -test.benchmem -test.benchtime 2m
goos: linux
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/query
cpu: Intel(R) Core(TM) i7-9850H CPU @ 2.60GHz
BenchmarkQuerySelect
BenchmarkQuerySelect/1000000SeriesWith1Samples
Creating 500000 1-sample series with 1ms interval in /tmp/testorbench-queryselect823049899/0
Creating 500000 1-sample series with 1ms interval in /tmp/testorbench-queryselect823049899/1
BenchmarkQuerySelect/1000000SeriesWith1Samples/select
BenchmarkQuerySelect/1000000SeriesWith1Samples/select-12  	      63	1955984103 ns/op	803750001 B/op	12500068 allocs/op
BenchmarkQuerySelect/100000SeriesWith100Samples
Creating 50000 50-sample series with 1ms interval in /tmp/testorbench-queryselect953795598/0
Creating 50000 50-sample series with 1ms interval in /tmp/testorbench-queryselect953795598/1
BenchmarkQuerySelect/100000SeriesWith100Samples/select
BenchmarkQuerySelect/100000SeriesWith100Samples/select-12 	     789	 184461340 ns/op	81817277 B/op	 1250058 allocs/op
BenchmarkQuerySelect/1SeriesWith10000000Samples
Creating 1 5000000-sample series with 1ms interval in /tmp/testorbench-queryselect150567957/0
Creating 1 5000000-sample series with 1ms interval in /tmp/testorbench-queryselect150567957/1
BenchmarkQuerySelect/1SeriesWith10000000Samples/select
BenchmarkQuerySelect/1SeriesWith10000000Samples/select-12 	    6386	  21871206 ns/op	18687412 B/op	  166722 allocs/op
PASS
```