@bwplotka 2021-06-28
This should be now fixed cc @jmichalek132  do you mind building from it and checking? 🤗 

In separate PRs I will try to optimize flow