@bwplotka 2021-06-28
While it works, at the current moment it introduces a lot of slow down:

```
benchstat -delta-test=none ../_dev/thanos/2021/select/no.txt ../_dev/thanos/2021/select/b.txt
name                                              old time/op    new time/op    delta
QuerySelect/1000000SeriesWith1Samples/select-12      1.85s ± 0%     2.48s ± 0%  +33.72%
QuerySelect/100000SeriesWith100Samples/select-12     176ms ± 0%     247ms ± 0%  +40.57%
QuerySelect/1SeriesWith10000000Samples/select-12    19.3ms ± 0%    26.3ms ± 0%  +36.24%

name                                              old alloc/op   new alloc/op   delta
QuerySelect/1000000SeriesWith1Samples/select-12      804MB ± 0%     804MB ± 0%   +0.00%
QuerySelect/100000SeriesWith100Samples/select-12    81.8MB ± 0%    81.8MB ± 0%   -0.00%
QuerySelect/1SeriesWith10000000Samples/select-12    18.7MB ± 0%    18.7MB ± 0%   +0.00%

name                                              old allocs/op  new allocs/op  delta
QuerySelect/1000000SeriesWith1Samples/select-12      12.5M ± 0%     12.5M ± 0%    0.00%
QuerySelect/100000SeriesWith100Samples/select-12     1.25M ± 0%     1.25M ± 0%    0.00%
QuerySelect/1SeriesWith10000000Samples/select-12      167k ± 0%      167k ± 0%    0.00%
```

To be investigated.