@bwplotka 2021-06-29
I transformed the whole thing into stream sort for efficiency, closing this one, stay tuned for next PR 🤗